Some refactoring and improvements.
- Change visibility.
- Packages fix.
