+++ src/main/java/com/glovoapp/backender/repositories/CourierRepository.java
@@ -34,7 +34,7 @@ public class CourierRepository {
                 .orElse(null);
     }
 
-    public List<Courier> findAll() {
+    List<Courier> findAll() {
         return new ArrayList<>(couriers);
     }
 }
