+++ src/test/java/com/glovoapp/backender/filters/OrderFilterTest.java
@@ -1,4 +1,4 @@
-package com.glovoapp.backender;
+package com.glovoapp.backender.filters;
 
 import com.glovoapp.backender.entities.Courier;
 import com.glovoapp.backender.entities.Location;
