+++ src/test/java/com/glovoapp/backender/repositories/CourierRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.glovoapp.backender;
+package com.glovoapp.backender.repositories;
 
 import com.glovoapp.backender.entities.Courier;
 import com.glovoapp.backender.entities.Location;
